feat: Looker 22.4 API updates (#1024)
* fix: 22_4 API

* fix: Additional fixes in unit tests for Hackathon

* fix: python integration tests for string id types

* fix: attempt to fix integration test for typescript

* fix TS SDK tests for new string IDs

* update python config notes, reformat python SDK files

* format python/looker_sdk/sdk/api{31,40}/methods.py

+++ kotlin/src/test/KotlinExample.kt
@@ -33,7 +33,7 @@ class KotlinExample {
             throw Error("No query found on ${tile.title}.")
         }
         try {
-            val queryId = tile.query_id!!.toLong()
+            val queryId = tile.query_id!!
             val task = sdk.ok<RenderTask>(sdk.create_query_render_task(queryId, format, 640, 480))
 
             if (task.id == null) {
